My attempt at rng with the investment/ stock options
I was having a lot of issue getting the right version of the game but I think i figured it out. So far I made numbers "bigger" so that they look more subnstantial. I also made it more clear that you can only select one option per week.  I put in the week adnvace object code I would try to use for if you make money or lose money on your stocks or company but it wont stop triggering if you like this code could you get it to where it only triggers once? if not its not neccessary and can be deleted.